# docs/api/xcdos_openapi.yaml
openapi: 3.1.0
info:
  title: XCDOS API
  version: 1.0.0
  description: >
    企业决策与执行操作系统 API。事实源（Gate 1）。
    契约对齐 ADR-0001~0005 + 评审 P0-01/02/03/04/06。
servers:
  - url: /api
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
    IfMatch:
      name: If-Match
      in: header
      required: false
      description: 乐观锁版本（P1-08），不匹配返回 409
      schema: { type: integer }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: 写接口幂等键（ADR-0005）
      schema: { type: string }
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
...
    Forecast:
      type: object
      required: [id, caseId, version, scenarios]
      properties:
        id: { type: string, format: uuid }
        caseId: { type: string, format: uuid }
        version: { type: integer }
        scenarios:
          type: array
          items:
            type: object
            properties:
              name: { type: string }
              probability: { type: number }
              outcome: { type: string }
              impact: { type: string }
              assumptions: { type: string }
        confidence: { type: number }
        modelSource: { type: string }
        agentRunId: { type: [string, "null"], format: uuid }
    Feedback:
      type: object
      required: [id, taskId, revision]
      properties:
        id: { type: string, format: uuid }
        taskId: { type: string, format: uuid }
        revision: { type: integer }
        supersededBy: { type: [string, "null"], format: uuid }
        qualityScore: { type: [number, "null"] }
        submittedAt: { type: string, format: date-time }
security:
  - bearerAuth: []
paths:
...
  /goals:
    post:
      operationId: createGoal
      responses: { '201': { description: 已创建 } }
    get:
      operationId: listGoals
      parameters:
        - { name: status, in: query, schema: { type: string } }
        - { name: goalId, in: query, schema: { type: string } }
      responses: { '200': { description: 列表 } }
  /goals/{id}:
    get:
      operationId: getGoal
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { '200': { description: 详情 }, '404': { description: 不存在 } }
    delete:
      operationId: deleteGoal
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        '204': { description: 已软删除 }
        '409': { description: GOAL_HAS_ACTIVE_TASKS, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
  /goals/{id}/status:
    patch:
      operationId: updateGoalStatus
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - $ref: '#/components/parameters/IfMatch'
      responses:
        '200': { description: 已更新 }
        '409': { description: VERSION_CONFLICT, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
  /decision-cases/{id}:
    get:
      operationId: getDecisionCase
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { '200': { description: 详情 } }
...
  /decision-cases/{id}/forecast:
    post:
      operationId: createForecast
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        '201': { description: 推演已持久化, content: { application/json: { schema: { $ref: '#/components/schemas/Forecast' } } } }
  /decision-cases/{id}/forecasts:
    get:
      operationId: listForecasts
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - { name: version, in: query, schema: { type: integer } }
      responses: { '200': { description: 推演版本列表 } }
...
  /tasks/{id}/feedback:
    post:
      operationId: submitFeedback
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '201': { description: 首次反馈, content: { application/json: { schema: { $ref: '#/components/schemas/Feedback' } } } }
        '409': { description: DUPLICATE_FEEDBACK（同日已有生效，改用 revisions）, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    get:
      operationId: getTaskFeedback
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { '200': { description: 反馈（含修订链） } }
  /feedbacks/{id}/revisions:
    post:
      operationId: reviseFeedback
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        '201': { description: 新修订, content: { application/json: { schema: { $ref: '#/components/schemas/Feedback' } } } }
        '403': { description: 非本人 }
        '409': { description: 已被更新修订替代 }
        '423': { description: 已锁定 }
...